requirements/requirements.txt
aioredis==1.3.1
anyio==3.6.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
blessed==1.19.1
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.2
django-cors-headers==3.13.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
Flask==2.2.2
globus-sdk==3.12.0
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
hubmap-commons==2.0.15
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
incremental==21.3.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonref==0.2
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.4
neo4j==5.0.1
networkx==2.8.7
pkgutil_resolve_name==1.3.10
property==2.2
prov==2.0.0
psutil==5.9.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
rdflib==6.2.0
redis==3.5.3
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.4.0
